Refactor: 게시글 작성일 포맷팅 유틸로 분리
- PostList.jsx 내부에 있던 날짜 포맷팅 메서드를 DateFormatter.js로 이동하여 재사용성 향상
